move feature declaration into the features enum
