+++ lib/flipper.rb
@@ -68,6 +68,7 @@ require 'flipper/adapter'
 require 'flipper/dsl'
 require 'flipper/errors'
 require 'flipper/feature'
+require 'flipper/features'
 require 'flipper/gate'
 require 'flipper/registry'
 require 'flipper/type'
